move viz things from mapbase to plotter
